fix - 메인이미지 삭제 시, null 확인

+++ src/main/java/com/project/tableforyou/domain/common/service/AssociatedEntityService.java
@@ -57,7 +57,9 @@ public class AssociatedEntityService {
         String mainImage = restaurantRepository.findMainImageById(restaurantId);
         List<String> menuImages = menuRepository.findMenuImagesByRestaurantId(restaurantId);
 
-        s3Service.deleteImage(mainImage);
+        if (mainImage != null)
+            s3Service.deleteImage(mainImage);
+
         images.forEach(s3Service::deleteImage);
         menuImages.forEach(s3Service::deleteImage);
